Initial import of the db_logger sources
These sources come almost-verbatim from the endbasic-service repository,
which at this point in time is not public.

The only changes I have made are to remove everything that did not belong
in this crate and to add license notices.

The original structure of the sources is preserved, as well as all
interfaces and dependencies.  A lot of work must be done to make this
simpler and actually usable by consumers, but it is a first good step to
capture the evolution of this code as its own standalone crate.